crash_nearmiss_analyzer/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.45.1
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
decord>=0.6.0
pillow>=9.0.0
scipy>=1.10.0
einops>=0.6.0
